# docs/swagger.yaml
definitions:
...
  controllers.changePasswordInput:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  controllers.userCreateInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_path:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  controllers.userUpdateInput:
    properties:
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      phone:
        type: string
      photo_path:
        type: string
      role:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_path:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /users:
    get:
      description: Get a list of User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get all User
      tags:
      - Admin
    post:
      description: Creating new User
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new User
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.userCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Create new User
      tags:
      - Admin
  /users/{id}:
    delete:
      description: Delete user by id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete user
      tags:
      - Admin
      - Buyer
      - Seller
    get:
      description: Get User by Id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get User detail
      tags:
      - Admin
      - Buyer
      - Seller
    put:
      description: Update User by id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.userUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Update User
      tags:
      - Admin
      - Buyer
      - Seller
...
  /users/change-password:
    post:
      description: Change Password
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to change password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.changePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Change Password
      tags:
      - Admin
      - Buyer
      - Seller
swagger: "2.0"